tabs: replace side nav with only top nav
