+++ packages/sidesail/lib/pages/tabs/zcash/zcash_melt_cast_page.dart
@@ -36,7 +36,7 @@ class ZCashMeltCastTabPage extends StatelessWidget {
         return SailPage(
           scrollable: true,
           widgetTitle: ZCashWidgetTitle(
-            depositNudgeAction: () => tabsRouter.setActiveIndex(ParentChainHome),
+            depositNudgeAction: () => tabsRouter.setActiveIndex(Tabs.ParentChainPeg.index),
           ),
           body: Padding(
             padding: const EdgeInsets.only(bottom: SailStyleValues.padding30),
