+++ packages/sidesail/lib/pages/tabs/zcash/zcash_shield_deshield_page.dart
@@ -33,7 +33,7 @@ class ZCashShieldDeshieldTabPage extends StatelessWidget {
         return SailPage(
           scrollable: true,
           widgetTitle: ZCashWidgetTitle(
-            depositNudgeAction: () => tabsRouter.setActiveIndex(ParentChainHome),
+            depositNudgeAction: () => tabsRouter.setActiveIndex(Tabs.ParentChainPeg.index),
           ),
           body: Padding(
             padding: const EdgeInsets.only(bottom: SailStyleValues.padding30),
