+++ packages/sidesail/lib/pages/tabs/zcash/zcash_transfer_page.dart
@@ -34,7 +34,7 @@ class ZCashTransferTabPage extends StatelessWidget {
         return SailPage(
           scrollable: true,
           widgetTitle: ZCashWidgetTitle(
-            depositNudgeAction: () => tabsRouter.setActiveIndex(ParentChainHome),
+            depositNudgeAction: () => tabsRouter.setActiveIndex(Tabs.ParentChainPeg.index),
           ),
           body: Padding(
             padding: const EdgeInsets.only(bottom: SailStyleValues.padding30),
