+++ packages/sidesail/lib/rpc/models/core_transaction.dart
@@ -5,7 +5,7 @@ import 'package:get_it/get_it.dart';
 import 'package:intl/intl.dart';
 import 'package:sail_ui/sail_ui.dart';
 import 'package:sidesail/bitcoin.dart';
-import 'package:sidesail/pages/tabs/dashboard_tab_page.dart';
+import 'package:sidesail/pages/tabs/sidechain_send_page.dart';
 import 'package:sidesail/rpc/rpc_sidechain.dart';
 
 class CoreTransaction {
